Refl probes bake null ref fix, Simplification
