--- Source/MotionBlurKeywordHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 94090c808a4fcd943b56d44f9bce4f89
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
